Bundle Update (finally), and lock gem versions where necessary
rails just should be updated, that was the impetus for this commit.

shoulda-matchers was locked because the next major version requires a
config block in rspec rails_helper.rb. will resolve in future commit.

I hadn't previously locked rottentomatoes and rspotify to a minor
version series, and I took the opportunity to do so now after updating
them. I depend on those for the APIs, and want to set myself up to
bundle update more freely without worrying about as many breaking
changes.