MovieGenre: correct uniqueness validation to scope to movie_id
Previously, it was checking that there are no rows that have the same
genre_id OR the same movie_id. That's incorrect, because I expect there
to be rows with the same movie_id but different genre_id's, and thus
also rows with the same genre_id, but different movie_id's.