SpotifyResource associations must have all keys converted to symbols
This is taken care of when you call .find on a SpotifyResource sub
class, since the #sanitize_result method is called, which ultimately
called #deep_symbolize_keys. This takes care of, for example, the nested
images hash from #query_result getting its keys converted to symbols.

The fact that those keys weren't symbols for Spotify::Artist#albums is
what prompted this bug fix. That happened because the Spotify::Albums
were not created via the .find mechanism, but instead by passing
attributes directly to #initialize, where we were only calling #symbolize_keys
even though those attributes were the same as those that would be
received from .find -> #query_api.

Creating SpotifyResource objects in this way can happen in several other
paths, such as Spotify::Song#artists. This commit will hopefully prevent
this bug from cropping up in those cases as well.

The fact that these classes can be created with or without calling #find
is what is causing this confusion, but always calling #find would result
in needless API requests, which is bad for this app. So this is probably
the best way to deal with this situation: ensuring that no matter what, #query_result
will always have symbol keys. Maybe there's a better way to express that
logical need.